MYRIAD-142 Specifying address:port instead of just port causes NumberFormatException if myriad.mapreduce.shuffle.port provided in any config file
In reviewing the code it was also discovered Myriad was only taking the first resource type offered instead of all resources of a type offered.  This meant if Myriad was running with role "Hadoop" and got an offer with cpu(Hadoop): 2 and cpu(*): 1, it would ignore the 1 cpu it could've taken from the * role.

This commit address Myriad-135, it is possible an offer may not contain a particular resource type, in which case an NPE occured.

In reviewing the code it was also discovered Myriad was only taking the first resource type offered instead of all resources of a type offered.  This meant if Myriad was running with role "Hadoop" and got an offer with cpu(Hadoop): 2 and cpu(*): 1, it would ignore the 1 cpu it could've taken from the * role.